move tests to tests folder
