+++ src_client/src/tests/Config.test.js
@@ -1,6 +1,6 @@
 import { fireEvent, screen, waitFor } from "@testing-library/react"
-import { server, rest, BASE_URL, renderApp,clearLocalStorage,renderAsFragment } from "../testing/utils"
-import Config from "./Config"
+import { server, rest, BASE_URL, renderApp,clearLocalStorage,renderAsFragment } from "../utils/testing"
+import Config from "../views/Config"
 
 beforeAll(() => server.listen())
 afterEach(() => server.resetHandlers())
