+++ src_client/src/tests/HelpCenter.test.js
@@ -1,7 +1,7 @@
 
 import { fireEvent, screen,waitFor } from "@testing-library/react"
-import { server, rest, BASE_URL, renderApp,clearLocalStorage,renderAsFragment } from "../testing/utils"
-import HelpCenter from "./HelpCentre"
+import { server, rest, BASE_URL, renderApp,clearLocalStorage,renderAsFragment } from "../utils/testing"
+import HelpCenter from "../views/HelpCentre"
 beforeAll(() => server.listen())
 afterEach(() => server.resetHandlers())
 afterAll(() => server.close())
